+++ microservice/src/main/java/com/api/microservice/service/UserService.java
@@ -3,6 +3,6 @@ package com.api.microservice.service;
 import com.api.microservice.models.User;
 
 public interface UserService {
-    User getUser(int id);
+    User getUser(Long id);
     User addUser(User u);
 }
